Séparation des données de jeu et amélioration des effets
