+++ Assets/Scripts/MonoBehavioursUsed/InventoryManager.cs
@@ -14,6 +14,20 @@ public class InventoryManager : MonoBehaviour
     private Dictionary<ItemData, int> itemUsesThisTurn = new();
     private Dictionary<ItemData, int> itemUsesThisBattle = new();
 
+    // --- Suivi des effets temporaires appliqués aux personnages ---
+    // Pour chaque personnage, on garde la liste des modificateurs actifs afin
+    // de pouvoir prolonger leur durée si nécessaire.
+    private class ActiveStatModifier
+    {
+        public BuffStatType stat;    // Statistique affectée
+        public float value;          // Valeur actuellement appliquée
+        public float remaining;      // Temps restant avant la fin de l'effet
+        public Coroutine routine;    // Coroutine responsable de la durée
+    }
+
+    // Dictionnaire principal : clé = unité concernée, valeur = liste de ses modificateurs
+    private readonly Dictionary<CharacterUnit, List<ActiveStatModifier>> activeModifiers = new();
+
     private void Awake()
     {
         if (Instance != null && Instance != this)
@@ -92,7 +106,7 @@ public class InventoryManager : MonoBehaviour
 
         float baseValue = GetBaseStat(target, stat);
         float value = isPercentage ? baseValue * amount / 100f : amount;
-        StartCoroutine(ApplyStatModifier(target, stat, value, duration));
+        ApplyStatModifier(target, stat, value, duration);
     }
 
     public void ApplyDebuff(CharacterUnit target, DebuffStatType stat, int amount, float duration, bool isPercentage)
@@ -102,14 +116,54 @@ public class InventoryManager : MonoBehaviour
 
         float baseValue = GetBaseStat(target, (BuffStatType)stat);
         float value = isPercentage ? baseValue * amount / 100f : amount;
-        StartCoroutine(ApplyStatModifier(target, (BuffStatType)stat, -value, duration));
+        ApplyStatModifier(target, (BuffStatType)stat, -value, duration);
     }
 
-    private IEnumerator ApplyStatModifier(CharacterUnit target, BuffStatType stat, float value, float duration)
+    /// <summary>
+    /// Applique ou prolonge un modificateur de statistique.
+    /// </summary>
+    private void ApplyStatModifier(CharacterUnit target, BuffStatType stat, float value, float duration)
     {
-        ModifyStat(target, stat, value);
-        yield return new WaitForSeconds(duration);
-        ModifyStat(target, stat, -value);
+        if (!activeModifiers.TryGetValue(target, out var list))
+        {
+            list = new List<ActiveStatModifier>();
+            activeModifiers[target] = list;
+        }
+
+        var modifier = list.Find(m => m.stat == stat);
+        if (modifier != null)
+        {
+            // Si un modificateur existe déjà, on cumule la valeur et on prolonge sa durée
+            ModifyStat(target, stat, value);
+            modifier.value += value;
+            modifier.remaining += duration;
+        }
+        else
+        {
+            modifier = new ActiveStatModifier
+            {
+                stat = stat,
+                value = value,
+                remaining = duration
+            };
+            ModifyStat(target, stat, value);
+            modifier.routine = StartCoroutine(StatModifierRoutine(target, modifier));
+            list.Add(modifier);
+        }
+    }
+
+    /// <summary>
+    /// Coroutine gérant la durée d'un modificateur.
+    /// </summary>
+    private IEnumerator StatModifierRoutine(CharacterUnit target, ActiveStatModifier modifier)
+    {
+        while (modifier.remaining > 0f)
+        {
+            yield return null;
+            modifier.remaining -= Time.deltaTime;
+        }
+        ModifyStat(target, modifier.stat, -modifier.value);
+        activeModifiers[target].Remove(modifier);
     }
 
     private void ModifyStat(CharacterUnit target, BuffStatType stat, float delta)
@@ -147,15 +201,26 @@ public class InventoryManager : MonoBehaviour
         target.interceptionImmunityTurns = Mathf.Max(target.interceptionImmunityTurns, turns);
     }
 
+    /// <summary>
+    /// Prolonge la durée de tous les effets temporaires actuellement actifs sur la cible.
+    /// </summary>
     public void ExtendEffectDurations(CharacterUnit target, int additionalTurns)
     {
         if (target == null || additionalTurns <= 0)
             return;
 
+        // Prolonge l'immunité à l'interception si présente
         if (target.interceptionImmunityTurns > 0)
             target.interceptionImmunityTurns += additionalTurns;
 
-        // TODO: étendre la durée des autres effets quand ils seront implémentés
+        // Prolonge également tous les buffs/debuffs suivis pour cette unité
+        if (activeModifiers.TryGetValue(target, out var list))
+        {
+            foreach (var modifier in list)
+            {
+                modifier.remaining += additionalTurns;
+            }
+        }
     }
 
     public void ApplySleep(CharacterUnit target)
