+++ Docs/NouveauxContenus.md
@@ -22,26 +22,26 @@ une utilisation illimitée.
   - Les joueurs expérimentés peuvent le combiner avec des MusicalMoves de protection pour sécuriser le retour au combat.
 
 
-## MusicalMove : Arp\xE8ge R\xE9g\xE9n\xE9rant
+## MusicalMove : Arpège Régénérant
 - **Type** : Soutien
-- **Effet** : soigne 20 PV \xE0 tous les alli\xE9s.
-- **Co\xFBt** : 2 points d'harmonie et 2 points de fatigue.
+- **Effet** : soigne 20 PV à tous les alliés.
+- **Coût** : 2 points d'harmonie et 2 points de fatigue.
 - **Gain** : aucun.
-- **Utilisation conseill\xE9e** : parfaite pour maintenir l'\xE9quipe en vie. Facile \xE0 ex\xE9cuter pour les d\xE9butants, mais permet aux joueurs exp\xE9riment\xE9s d'encha\xEEner de longs combats.
+- **Utilisation conseillée** : parfaite pour maintenir l'équipe en vie. Facile à exécuter pour les débutants, mais permet aux joueurs expérimentés d'enchaîner de longs combats.
 
-## MusicalMove : Staccato \xC9tourdissant
-- **Type** : D\xE9buff
+## MusicalMove : Staccato Étourdissant
+- **Type** : Débuff
 - **Effet** : endort une cible ennemie pendant un tour.
-- **Co\xFBt** : 2 points d'harmonie et 2 points de fatigue.
+- **Coût** : 2 points d'harmonie et 2 points de fatigue.
 - **Gain** : aucun.
-- **Utilisation conseill\xE9e** : id\xE9al pour interrompre un adversaire puissant le temps de pr\xE9parer une combinaison.
+- **Utilisation conseillée** : idéal pour interrompre un adversaire puissant le temps de préparer une combinaison.
 
-## Item : Cl\xE9 du Courage
-- **Effet** : augmente la Force d'un alli\xE9 de 10 pendant deux tours.
-- **Co\xFBt** : aucun.
-- **Utilisation conseill\xE9e** : utile d\xE8s les premiers combats pour booster un personnage offensif. Les strat\xE8ges aguerris l'exploiteront pour maximiser les d\xE9g\xE2ts d'un encha\xEEnement de MusicalMoves.
+## Item : Clé du Courage
+- **Effet** : augmente la Force d'un allié de 10 pendant deux tours.
+- **Coût** : aucun.
+- **Utilisation conseillée** : utile dès les premiers combats pour booster un personnage offensif. Les stratèges aguerris l'exploiteront pour maximiser les dégâts d'un enchaînement de MusicalMoves.
 
 ## Item : Bombe Sonique
-- **Effet** : inflige 15 points de d\xE9g\xE2ts \xE0 tous les ennemis.
-- **Co\xFBt** : aucun.
-- **Utilisation conseill\xE9e** : permet d'affaiblir tout un groupe d'ennemis avant d'utiliser des attaques cibl\xE9es plus puissantes.
+- **Effet** : inflige 15 points de dégâts à tous les ennemis.
+- **Coût** : aucun.
+- **Utilisation conseillée** : permet d'affaiblir tout un groupe d'ennemis avant d'utiliser des attaques ciblées plus puissantes.
